Create README.md
# Stock-Volatility-Prediction
The goal of this program is to read the stock prices (originally given in one minute intervals for a period of a year) from a csv file, clean and process them, compute the volatility of each stock in a given time interval (e.g. in a day), and finally use the ARIMA model to predict the future volatility of each stock (originally for a period of a month).
Obviously the steps needed for processing the data, or choosing the model (to fit the data and make predictions) depend on factors such as the format and existing deficiencies of the data and what is expected of the program, as each data set comes with its own unique needs for different methods needed to clean and process the data.
Any constructive criticism or suggestion on ways to improve this program, features that could be added to make better predictions, or even using more suitable models is greatly appreciated.
This program was written using Jupyter Notebook 6.0.1.